Move dashboard screen to a separate file. Fix a typo
