+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:fast_qr_reader_view/fast_qr_reader_view.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:medication_book/ui/screen/spash_screen.dart';
+import 'package:medication_book/ui/screen/splash_screen.dart';
 // import 'package:permission_handler/permission_handler.dart';
 
 List<CameraDescription> cameras;
@@ -25,7 +25,6 @@ Future<Null> main() async {
     runApp(MyApp());
   });
 }
-
 
 class MyApp extends StatelessWidget {
   // This widget is the root of your application.
